+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New features
 
+* New cop `ExtraSpacing` points out unnecessary spacing in files. ([@blainesch][])
 * New cop `EmptyLinesAroundBlockBody` provides same functionality as the EmptyLinesAround(Class|Method|Module)Body but for blocks. ([@jcarbo][])
 * New cop `Style/EmptyElse` checks for empty `else`-clauses. ([@Koronen][])
 
@@ -1178,3 +1179,4 @@
 [@oneamtu]: https://github.com/oneamtu
 [@toy]: https://github.com/toy
 [@Koronen]: https://github.com/Koronen
+[@blainesch]: https://github.com/blainesch